+++ node/src/GlideClient.ts
@@ -223,13 +223,14 @@ export class GlideClient extends BaseClient {
         });
     }
 
-    /** Ping the Redis server.
+    /**
+     * Pings the Redis server.
+     *
      * @see {@link https://valkey.io/commands/ping/|valkey.io} for details.
      *
-     * @param message - An optional message to include in the PING command.
-     * If not provided, the server will respond with "PONG".
-     * If provided, the server will respond with a copy of the message.
-     * @param decoder - (Optional) {@link Decoder} type which defines how to handle the response. If not set, the default decoder from the client config will be used.
+     * @param message - (Optional) A message to include in the PING command.
+     *     If not provided, the server will respond with "PONG".
+     *     If provided, the server will respond with a copy of the message.
      * @returns - "PONG" if `message` is not provided, otherwise return a copy of `message`.
      *
      * @example
@@ -246,12 +247,12 @@ export class GlideClient extends BaseClient {
      * console.log(result); // Output: 'Hello'
      * ```
      */
-    public async ping(options?: {
-        message?: GlideString;
-        decoder?: Decoder;
-    }): Promise<GlideString> {
-        return this.createWritePromise(createPing(options?.message), {
-            decoder: options?.decoder,
+    public async ping(message?: GlideString): Promise<GlideString> {
+        return this.createWritePromise(createPing(message), {
+            decoder:
+                !message || typeof message === "string"
+                    ? Decoder.String
+                    : Decoder.Bytes,
         });
     }
 
@@ -266,11 +267,13 @@ export class GlideClient extends BaseClient {
         return this.createWritePromise(createInfo(options));
     }
 
-    /** Change the currently selected Redis database.
+    /**
+     * Changes the currently selected Redis database.
+     *
      * @see {@link https://valkey.io/commands/select/|valkey.io} for details.
      *
      * @param index - The index of the database to select.
-     * @returns A simple OK response.
+     * @returns A simple `"OK"` response.
      *
      * @example
      * ```typescript
@@ -280,13 +283,19 @@ export class GlideClient extends BaseClient {
      * ```
      */
     public async select(index: number): Promise<"OK"> {
-        return this.createWritePromise(createSelect(index));
+        return this.createWritePromise(createSelect(index), {
+            decoder: Decoder.String,
+        });
     }
 
-    /** Get the name of the primary's connection.
+    /**
+     * Gets the name of the primary's connection.
+     *
      * @see {@link https://valkey.io/commands/client-getname/|valkey.io} for more details.
      *
-     * @returns the name of the client connection as a string if a name is set, or null if no name is assigned.
+     * @param decoder - (Optional) {@link Decoder} type which defines how to handle the response.
+     *     If not set, the {@link BaseClientConfiguration.defaultDecoder|default decoder} will be used.
+     * @returns The name of the client connection as a string if a name is set, or `null` if no name is assigned.
      *
      * @example
      * ```typescript
@@ -295,8 +304,10 @@ export class GlideClient extends BaseClient {
      * console.log(result); // Output: 'Client Name'
      * ```
      */
-    public async clientGetName(): Promise<string | null> {
-        return this.createWritePromise(createClientGetName());
+    public async clientGetName(decoder?: Decoder): Promise<GlideString | null> {
+        return this.createWritePromise(createClientGetName(), {
+            decoder: decoder,
+        });
     }
 
     /** Rewrite the configuration file with the current configuration.
@@ -332,10 +343,12 @@ export class GlideClient extends BaseClient {
         return this.createWritePromise(createConfigResetStat());
     }
 
-    /** Returns the current connection id.
+    /**
+     * Returns the current connection ID.
+     *
      * @see {@link https://valkey.io/commands/client-id/|valkey.io} for details.
      *
-     * @returns the id of the client.
+     * @returns The ID of the client.
      */
     public async clientId(): Promise<number> {
         return this.createWritePromise(createClientId());
@@ -380,7 +393,9 @@ export class GlideClient extends BaseClient {
         return this.createWritePromise(createConfigSet(parameters));
     }
 
-    /** Echoes the provided `message` back.
+    /**
+     * Echoes the provided `message` back.
+     *
      * @see {@link https://valkey.io/commands/echo|valkey.io} for more details.
      *
      * @param message - The message to be echoed back.
@@ -393,8 +408,11 @@ export class GlideClient extends BaseClient {
      * console.log(echoedMessage); // Output: 'valkey-glide'
      * ```
      */
-    public async echo(message: string): Promise<string> {
-        return this.createWritePromise(createEcho(message));
+    public async echo(message: GlideString): Promise<GlideString> {
+        return this.createWritePromise(createEcho(message), {
+            decoder:
+                typeof message === "string" ? Decoder.String : Decoder.Bytes,
+        });
     }
 
     /** Returns the server time
@@ -925,6 +943,8 @@ export class GlideClient extends BaseClient {
      * ```
      */
     public async unwatch(): Promise<"OK"> {
-        return this.createWritePromise(createUnWatch());
+        return this.createWritePromise(createUnWatch(), {
+            decoder: Decoder.String,
+        });
     }
 }