+++ node/tests/SharedTests.ts
@@ -161,7 +161,19 @@ export function runBaseTests(config: {
         async (protocol) => {
             await runTest(
                 async (client: BaseClient) => {
-                    expect(await client.clientGetName()).toBe("TEST_CLIENT");
+                    expect(await client.clientGetName()).toEqual("TEST_CLIENT");
+
+                    if (client instanceof GlideClient) {
+                        expect(
+                            await client.clientGetName(Decoder.Bytes),
+                        ).toEqual(Buffer.from("TEST_CLIENT"));
+                    } else {
+                        expect(
+                            await client.clientGetName({
+                                decoder: Decoder.Bytes,
+                            }),
+                        ).toEqual(Buffer.from("TEST_CLIENT"));
+                    }
                 },
                 protocol,
                 { clientName: "TEST_CLIENT" },
@@ -487,20 +499,17 @@ export function runBaseTests(config: {
         async (protocol) => {
             await runTest(async (client: BaseClient) => {
                 const pongEncoded = Buffer.from("PONG");
-                const helloEncoded = Buffer.from("Hello");
                 expect(await client.ping()).toEqual("PONG");
-                expect(await client.ping({ message: "Hello" })).toEqual(
-                    "Hello",
-                );
-                expect(await client.ping({ decoder: Decoder.Bytes })).toEqual(
-                    pongEncoded,
-                );
                 expect(
-                    await client.ping({
-                        message: "Hello",
-                        decoder: Decoder.Bytes,
-                    }),
-                ).toEqual(helloEncoded);
+                    client instanceof GlideClient
+                        ? await client.ping("Hello")
+                        : await client.ping({ message: "Hello" }),
+                ).toEqual("Hello");
+                expect(
+                    client instanceof GlideClient
+                        ? await client.ping(pongEncoded)
+                        : await client.ping({ message: pongEncoded }),
+                ).toEqual(pongEncoded);
             }, protocol);
         },
         config.timeout,
@@ -4808,6 +4817,9 @@ export function runBaseTests(config: {
             await runTest(async (client: BaseClient) => {
                 const message = uuidv4();
                 expect(await client.echo(message)).toEqual(message);
+                expect(await client.echo(Buffer.from(message))).toEqual(
+                    Buffer.from(message),
+                );
             }, protocol);
         },
         config.timeout,